+++ src/components/ProfileDetail.js
@@ -1,24 +1,72 @@
+// @flow
+
 import React, {Component} from 'react';
-import {TouchableOpacity, Alert, View, Text} from 'react-native';
+import {StyleSheet, TouchableOpacity, Alert, Image, View, Text} from 'react-native';
 import Screen from './Screen';
-import ProfileInfo from './ProfileInfo';
 import ProfileAvatar from './ProfileAvatar';
 import Card from './Card';
-import CellWithText from './CellWithText';
-import Separator from './Separator';
-import friendStore from '../store/friendStore';
 import Profile from '../model/Profile';
-import model from '../model/model';
-import message from '../store/messageStore';
 import location from '../store/locationStore';
-import statem, {ProfileDetailsState} from '../../gen/state';
 import profileStore from '../store/profileStore';
+import friendStore from '../store/friendStore';
 import {observer} from 'mobx-react/native';
+import {observable} from 'mobx';
+import Bots from '../model/Bots';
 import {k} from './Global';
-import {navBarTextColorDay, navBarTextColorNight} from '../globals';
+import {colors} from '../constants';
+import botStore from '../store/botStore';
+import NavBar from './NavBar';
+import NavTitle from './NavTitle';
+import NavBarRightButton from './NavBarRightButton';
+import BotListView from './BotListView';
+import autobind from 'autobind-decorator';
+import BotButton from './BotButton';
+import statem from '../../gen/state';
+import messageStore from '../store/messageStore';
 
+const Separator = () => <View style={{width: 1 * k, top: 7 * k, height: 34 * k, backgroundColor: colors.SILVER}} />;
+
+type Props = {
+    item: Object
+};
+
+const Header = ({profile, isDay}: {profile: Profile, isDay: boolean}) => (
+    <View style={{backgroundColor: colors.WHITE}}>
+        <Card style={styles.header}>
+            <ProfileAvatar size={100} isDay={isDay} profile={profile} tappable={false} />
+            <Text style={styles.displayName}>{profile.displayName}</Text>
+            <Text style={styles.tagline}>{profile.tagline}</Text>
+            <View style={styles.metabar}>
+                <View style={{flex: 1}}>
+                    <Text style={styles.number}>123</Text>
+                    <Text style={styles.word}>BOTS</Text>
+                </View>
+                <Separator />
+                <View style={{flex: 1}}>
+                    <Text style={styles.number}>123</Text>
+                    <Text style={styles.word}>FOLLOWERS</Text>
+                </View>
+                <Separator />
+                <View style={{flex: 1}}>
+                    <Text style={styles.number}>123</Text>
+                    <Text style={styles.word}>FOLLOWING</Text>
+                </View>
+            </View>
+        </Card>
+        {profile.isFollowed &&
+            <View style={{height: 15 * k}}>
+                <TouchableOpacity onPress={this.unfollow} style={{position: 'absolute', left: 120 * k, bottom: 10 * k}}>
+                    <Image source={require('../../images/buttonFollowing.png')} />
+                </TouchableOpacity>
+            </View>}
+    </View>
+);
+
+@autobind
 @observer
 export default class ProfileDetail extends Component {
+    @observable bots = new Bots();
+    props: Props;
     // static onRight({item, title}) {
     //   Actions.profileOptions({item, title});
     // }
@@ -26,108 +74,49 @@ export default class ProfileDetail extends Component {
         return <Text>{item.firstName} {item.lastName}</Text>;
     }
 
+    async unfollow() {
+        const profile: Profile = profileStore.create(this.props.item);
+        Alert.alert(null, `Are you sure you want to unfollow ${profile.handle}?`, [
+            {text: 'Cancel', style: 'cancel'},
+            {
+                text: 'Unfollow',
+                style: 'destructive',
+                onPress: () => friendStore.unfollow(profile),
+            },
+        ]);
+    }
+
+    async componentDidMount() {
+        if (this.props.item) {
+            await botStore.list(this.bots, this.props.item);
+        }
+    }
+
     render() {
         const isDay = location.isDay;
         const profile: Profile = profileStore.create(this.props.item);
-        const state: ProfileDetailsState = statem.profileDetails;
         return (
-            <Screen isDay={isDay} style={{paddingTop: 70 * k}}>
-                <View>
-                    <ProfileAvatar isDay={isDay} profile={profile} tappable={false} />
-                    <ProfileInfo isDay={isDay} profile={profile} message={message} />
-                    {!profile.isOwn &&
-                        <Card isDay={isDay} style={{opacity: 0.95}}>
-                            <View style={{padding: 15 * k}}>
-                                <Text
-                                    style={{
-                                        fontFamily: 'Roboto-Medium',
-                                        fontSize: 16,
-                                        color: isDay ? navBarTextColorDay : navBarTextColorNight,
-                                    }}
-                                >
-                                    Options
-                                </Text>
-                            </View>
-                            <Separator width={1} />
-                            {profile.isFollowed &&
-                                profile.isFollower &&
-                                <View>
-                                    <TouchableOpacity onPress={() => setTimeout(() => state.openPrivateChat(profile))}>
-                                        <CellWithText isDay={isDay}>Send a message</CellWithText>
-                                    </TouchableOpacity><Separator width={1} />
-                                </View>}
-                            {profile.isFollowed &&
-                                <View>
-                                    <TouchableOpacity
-                                        onPress={() =>
-                                            Alert.alert('Are you sure?', null, [
-                                                {
-                                                    text: 'Yes',
-                                                    onPress: () => friendStore.unfollow(profile),
-                                                },
-                                                {text: 'No'},
-                                            ])}
-                                    >
-                                        <CellWithText isDay={isDay}>
-                                            Unfollow {profile.displayName}
-                                        </CellWithText>
-                                    </TouchableOpacity><Separator width={1} />
-                                </View>}
-                            {!profile.isFollowed &&
-                                <View>
-                                    <TouchableOpacity onPress={() => friendStore.add(profile)}>
-                                        <CellWithText isDay={isDay}>
-                                            Follow {profile.displayName}
-                                        </CellWithText>
-                                    </TouchableOpacity><Separator width={1} />
-                                </View>}
-
-                            {profile.hidePosts &&
-                                <View>
-                                    <TouchableOpacity onPress={() => state.showPosts(profile)}>
-                                        <CellWithText image={require('../../images/show.png')} isDay={isDay}>
-                                            Show {profile.displayName}'s
-                                            Posts
-                                        </CellWithText>
-                                    </TouchableOpacity><Separator width={1} />
-                                </View>}
-                            {!profile.isFollowed &&
-                                !profile.hidePosts &&
-                                <View>
-                                    <TouchableOpacity onPress={() => state.hidePosts(profile)}>
-                                        <CellWithText image={require('../../images/hide.png')} isDay={isDay}>
-                                            Hide {profile.displayName}'s
-                                            Posts
-                                        </CellWithText>
-                                    </TouchableOpacity><Separator width={1} />
-                                </View>}
-
-                            {!profile.isFollower &&
-                                !profile.isFollowed &&
-                                !profile.isBlocked &&
-                                <TouchableOpacity
-                                    onPress={() =>
-                                        Alert.alert('Are you sure?', null, [
-                                            {
-                                                text: 'Yes',
-                                                onPress: () => friendStore.block(profile),
-                                            },
-                                            {text: 'No'},
-                                        ])}
-                                >
-                                    <CellWithText isDay={isDay} textStyle={{color: 'red'}}>
-                                        Block {profile.firstName || profile.displayName}
-                                    </CellWithText>
-                                </TouchableOpacity>}
-
-                            {profile.isBlocked &&
-                                <TouchableOpacity onPress={() => friendStore.unblock(profile)}>
-                                    <CellWithText isDay={isDay} textStyle={{color: 'red'}}>
-                                        Unblock {profile.firstName || profile.displayName}
-                                    </CellWithText>
-                                </TouchableOpacity>}
-                        </Card>}
-                </View>
+            <Screen isDay={isDay}>
+                <BotListView
+                    ref='list'
+                    list={this.bots}
+                    user={this.props.item}
+                    hideAvatar
+                    header={() => <Header profile={profile} isDay={isDay} />}
+                />
+                <NavBar>
+                    <NavTitle onPress={() => this.refs.list.scrollToTop()}>@{profile.handle}</NavTitle>
+                    {profile.isOwn && <NavBarRightButton active><Image source={require('../../images/settings.png')} /></NavBarRightButton>}
+                    {profile.isMutual &&
+                        <NavBarRightButton onPress={() => statem.profileDetails.openPrivateChat({item: messageStore.createChat(profile).id})} active>
+                            <Image source={require('../../images/createmessage.png')} />
+                        </NavBarRightButton>}
+                    {!profile.isFollowed &&
+                        <NavBarRightButton onPress={() => friendStore.follow(profile)} active>
+                            <Text style={styles.follow}>Follow</Text>
+                        </NavBarRightButton>}
+                </NavBar>
+                <BotButton />
             </Screen>
         );
     }
@@ -136,3 +125,47 @@ export default class ProfileDetail extends Component {
 ProfileDetail.propTypes = {
     item: React.PropTypes.any.isRequired,
 };
+
+const styles = StyleSheet.create({
+    header: {
+        paddingLeft: 0,
+        paddingRight: 0,
+        // @TODO: extract header padding to a global?
+        paddingTop: 62 * k,
+    },
+    displayName: {
+        paddingTop: 10 * k,
+        fontFamily: 'Roboto-Regular',
+        fontSize: 16 * k,
+        color: colors.navBarTextColorDay,
+        textAlign: 'center',
+    },
+    tagline: {
+        paddingBottom: 23 * k,
+        fontFamily: 'Roboto-Regular',
+        fontSize: 13 * k,
+        color: colors.navBarTextColorDay,
+        textAlign: 'center',
+    },
+    metabar: {
+        flexDirection: 'row',
+        paddingBottom: 30 * k,
+    },
+    number: {
+        fontFamily: 'Roboto-Regular',
+        fontSize: 22 * k,
+        color: colors.navBarTextColorDay,
+        textAlign: 'center',
+    },
+    follow: {
+        color: colors.PINK,
+        fontFamily: 'Roboto-Regular',
+        fontSize: 15 * k,
+    },
+    word: {
+        fontFamily: 'Roboto-Light',
+        fontSize: 11 * k,
+        color: colors.DARK_GREY,
+        textAlign: 'center',
+    },
+});